requirements.txt
sentry-sdk>=1.0.0,<2.0.0
tailhead>=1.0.2
pyyaml~=6.0